Because implementations of IConfigurationService are specific to a single application, and to ensure that it is impossible to register the wrong one during Startup, move the implementations to their respective projects.
